+++ packages/atomic/src/components/common/atomic-modal/atomic-modal.tsx
@@ -54,6 +54,10 @@ export class AtomicModal implements InitializableComponent<AnyBindings> {
   @Prop({reflect: true, mutable: true}) isOpen = false;
   @Prop({mutable: true}) close: () => void = () => (this.isOpen = false);
   @Prop({reflect: true}) noFocusTrap = false;
+  /**
+   * Whether to display the open and close animations over the entire page or the atomic-modal only.
+   */
+  @Prop({reflect: true}) boundary: 'page' | 'element' = 'page';
 
   @Event() animationEnded!: EventEmitter<never>;
 
@@ -134,9 +138,9 @@ export class AtomicModal implements InitializableComponent<AnyBindings> {
     const Content = () => (
       <article
         part="container"
-        class={`flex flex-col justify-between bg-background text-on-background ${
-          this.isOpen ? 'animate-scaleUpModal' : 'animate-slideDownModal'
-        } ${this.wasEverOpened ? '' : 'invisible'}`}
+        class={`flex flex-col justify-between bg-background text-on-background
+          ${this.isOpen ? 'animate-scaleUpModal' : 'animate-slideDownModal'}
+          ${this.wasEverOpened ? '' : 'invisible'}`}
         onAnimationEnd={() => this.animationEnded.emit()}
         ref={(ref) => (this.animatableContainer = ref)}
       >
@@ -183,7 +187,10 @@ export class AtomicModal implements InitializableComponent<AnyBindings> {
       <Host class={this.getClasses().join(' ')}>
         <div
           part="backdrop"
-          class="fixed left-0 top-0 right-0 bottom-0 z-[9999]"
+          class={`
+            ${this.boundary === 'page' ? 'fixed' : 'absolute'}
+            left-0 top-0 right-0 bottom-0 z-[9999]
+          `}
           onClick={(e) => e.target === e.currentTarget && this.close()}
         >
           {this.noFocusTrap ? (