+++ packages/headless/src/features/instant-results/instant-results-slice.test.ts
@@ -10,11 +10,24 @@ import {InstantResultCache} from './instant-results-state';
 const id1 = 'search_box_1';
 const id2 = 'search_box_2';
 
-const getSearchBoxInstantResultsState = (id: string, q = '', cache = {}) => ({
+const getSearchBoxInstantResultsState = (
+  id: string,
+  q = '',
+  cache: Record<string, InstantResultCache> = {}
+) => ({
   [id]: {
     q,
     cache,
   },
+});
+
+const initialEmptyCache: () => InstantResultCache = () => ({
+  isLoading: true,
+  error: null,
+  results: [],
+  expiresAt: 0,
+  isActive: false,
+  searchUid: '',
 });
 
 describe('instant results slice', () => {
@@ -41,7 +54,7 @@ describe('instant results slice', () => {
     it('does not override an existing search box', () => {
       const searchBox1 = () =>
         getSearchBoxInstantResultsState(id1, 'some_query', {
-          some_query: {isLoading: true, error: null, results: []},
+          some_query: initialEmptyCache(),
         });
 
       expect(
@@ -89,7 +102,7 @@ describe('instant results slice', () => {
 
         const initialState = getSearchBoxInstantResultsState(id1, query);
         const expectedState = getSearchBoxInstantResultsState(id1, query, {
-          some_query: {isLoading: true, error: null, results: [], expiresAt: 0},
+          some_query: initialEmptyCache(),
         });
 
         expect(instantResultsReducer(initialState, action)).toEqual(
@@ -106,10 +119,10 @@ describe('instant results slice', () => {
         });
 
         const initialState = getSearchBoxInstantResultsState(id1, query, {
-          some_query: {isLoading: false, error: null, results: []},
+          some_query: initialEmptyCache(),
         });
         const expectedState = getSearchBoxInstantResultsState(id1, query, {
-          some_query: {isLoading: true, error: null, results: []},
+          some_query: initialEmptyCache(),
         });
 
         expect(instantResultsReducer(initialState, action)).toEqual(
@@ -131,12 +144,7 @@ describe('instant results slice', () => {
         };
         const expectedState = {
           ...getSearchBoxInstantResultsState(id1, query, {
-            some_query: {
-              isLoading: true,
-              error: null,
-              results: [],
-              expiresAt: 0,
-            },
+            some_query: initialEmptyCache(),
           }),
           ...getSearchBoxInstantResultsState(id2, query),
         };
@@ -157,7 +165,7 @@ describe('instant results slice', () => {
       it('updates results in correct searchbox and query cache', () => {
         const query = 'some_query';
         const action = fetchInstantResults.fulfilled(
-          {results: [buildMockResult()]},
+          {results: [buildMockResult()], searchUid: 'someid'},
           'req_id',
           {
             id: id1,
@@ -169,25 +177,22 @@ describe('instant results slice', () => {
         const makeState = (some_query: InstantResultCache) => ({
           ...getSearchBoxInstantResultsState(id1, query, {
             some_query,
-            some_other_query: {isLoading: false, error: null, results: []},
+            some_other_query: initialEmptyCache(),
           }),
           ...getSearchBoxInstantResultsState(id2, query, {
-            some_query: {isLoading: false, error: null, results: []},
+            some_query: initialEmptyCache(),
           }),
         });
 
-        const initialState = makeState({
-          isLoading: true,
-          error: null,
-          results: [],
-          expiresAt: 0,
-        });
+        const initialState = makeState(initialEmptyCache());
 
         const expectedState = makeState({
           isLoading: false,
           error: null,
           results: [buildMockResult()],
           expiresAt: 0,
+          isActive: true,
+          searchUid: 'someid',
         });
 
         expect(instantResultsReducer(initialState, action)).toEqual(
@@ -197,7 +202,7 @@ describe('instant results slice', () => {
       it('sets correct isLoading, error and expiresAt properties', () => {
         const query = 'some_query';
         const action = fetchInstantResults.fulfilled(
-          {results: [buildMockResult()]},
+          {results: [buildMockResult()], searchUid: 'someid'},
           'req_id',
           {
             id: id1,
@@ -210,21 +215,18 @@ describe('instant results slice', () => {
         const makeState = (some_query: InstantResultCache) =>
           getSearchBoxInstantResultsState(id1, query, {
             some_query,
-            some_other_query: {isLoading: false, error: null, results: []},
+            some_other_query: initialEmptyCache(),
           });
 
-        const initialState = makeState({
-          isLoading: true,
-          error: null,
-          results: [],
-          expiresAt: 0,
-        });
+        const initialState = makeState(initialEmptyCache());
 
         const expectedState = makeState({
           isLoading: false,
           error: null,
           results: [buildMockResult()],
           expiresAt: Date.now() + 10000,
+          isActive: true,
+          searchUid: 'someid',
         });
 
         expect(instantResultsReducer(initialState, action)).toEqual(