+++ packages/samples/headless-react/src/utils/filter-protocol.ts
@@ -1,6 +1,6 @@
 export function filterProtocol(uri: string) {
   // Filters out dangerous URIs that can create XSS attacks such as `javascript:`.
-  const isAbsolute = /^(https?|ftp|file|mailto|tel):/i.test(uri);
+  const isAbsolute = /^(https?|ftp|file|mailto|tel|sip):/i.test(uri);
   const isRelative = /^(\/|\.\/|\.\.\/)/.test(uri);
 
   return isAbsolute || isRelative ? uri : '';